Don't mutate `self` in `save_changes`
With the new behavior of `Option` in `changeset_for`, we can no longer
assume that a type which implements `changeset_for` can also implement
`Queriable`. I was unhappy with the API of `save_changes` anyway, as it
seemed to have limited usefulness. Since I don't have a great way to
conditionally implement the old form, I've simply changed it to
something that we know will work for all types. If possible in the
future, I might still like to default this type parameter to `Self`.

Fixes #50