Release 0.3.0
This is a bit of a smaller release, but I will be going on vacation for
10 days tomorrow and wanted to get some of this out before I leave.
This release includes several great quality of life changes:

- `#[changeset_for]` now treats `Option` as an optional update.
  Previously, `None` would indicate inserting NULL into a field. `None`
  will now be skipped in the update, better mirroring the behavior of most
  APIs.
- `infer_schema!` and `infer_table_from_schema!` have been added, so you
  no longer have to manually write `table!` all over the place
- Connection pooling support has been added via R2D2

This will likely be the last *super short* release, and we will probably
be biweekly from now on unless there's a specific reason to break that.
See you all in 10 days! 🚢

[ci skip]