Use bind params for managing __schema_migrations
Interestingly enough, the easiest way to accomplish this is to just use
our own DSLs. We can't use codegen here, as that would create a circular
dependency, but we just need to implement `Insertable` by hand.

There were a few strange quirks that came as a result of this. I had to
move `macros` up to the top so it could be used in `migrations`, which
meant it could not do `use $crate::*;`. Additionally, we once again run
into the bug where we can only use traits in public modules in our own
crate.