Fix `#[changeset_for]` with single field structs
This error occurred because we were always assuming the type and body to
work with tuples, and `Changeset` isn't implemented for `(A)`. We do the
same workaround in the generation for `insertable_into`. I'd rather just
add `(A)` to the tuple impls, but this is unfortunately impossible due
to https://github.com/rust-lang/rust/issues/30762