Infer schema properly when array columns are used
Note: We don't ever actually use the added column, because we don't need
to. Its presence alone is enough to cause the tests to fail.

This implementation is a hack. It's awful. I need to write a more
extensible system for type lookups. But right now I don't care. This is
blocking 0.4, and I can fix it in 0.5.

Fixes #65