Add support for `diesel migration generate migration_name` to CLI
This command will create a timestamped directory in the migrations
folder, along with the empty files `up.sql` and `down.sql`. Right now
I'm spitting out absolute paths, it would be nice to print them as
relative instead (unsure if this should be relative to the cwd or to the
parent of the migrations directory).