Add a bare `select` function for select statements with no from clause
In addition to this function, I have also added a `sql` function, which
is shorthand for `SqlLiteral` (but only takes `&str`, not `String`), and
`in_time_zone` which is not part of the public API.

`sql` takes `&str`, instead of `T: Into<String>`, because the first type
parameter is always going to need to be passed, and I find
`sql::<Timestamp, _>` ugly. There's always `SqlLiteral::new` if you need
to pass a `String`, or you can just deref it to `str` and accept the
copying.

I had hoped to eliminate `query_sql_params`, but there's one test we
have that's using it which requires a where clause. I can't just hack it
and do `select(sql("1 WHERE ..."))` because it uses bind params. I'll
implement `FilterDsl` for `BareSelectStatement` at some point. I'd also
like to remove `query_sql`, as I think that `select(sql` covers it
suffciently.

I've opted to leave `in_time_zone` out of the public API, as it should
actually have the return type of `Timestamptz`, and I'm really unsure
how various infix predicates with the lhs being `Timestamp` and the rhs
being `Timestamptz` are going to interact. We might be able to get away
with just not allowing it and have it be a non-issue, similarly to
`VarChar` and `Text`, but it needs further exploration.

Fixes #97.