Remove use of `aster` from `queryable` in `codegen`
This is the first piece of several commits which will aim to remove the
usage of `aster` and `quasi` entirely. Our actual breakage on nightly
tends to be very rare, however these two crate break almost weekly. I'm
hoping that removing their usage will help to reduce the amount of
update burden we have to keep up with.