Remove `Expression#to_insert_sql`
This is part of a larger refactoring which will be coming in many steps.
This piece is a prerequisite for basically everything else.

`to_insert_sql` has been a general wart in our implementation that was a
side effect of how we used to structure things. Ultimately the only way
that it's really used is to replace null bind params with `DEFAULT` on
inserts.

This cleans up our code a good bit, and starts to allow us to separate
`Expression` and `QueryFragment` into more decoupled entities, which
will be important going forward.