Remove `QueryBuilder` related methods from `UpdateTarget`
This is part of an ongoing refactoring to remove all methods which take
a `&mut QueryBuilder` other than those on `QueryFragment`. The goal is
to reduce the number of places that need to know about conversion to
SQL, thus reducing the number of things that need to become generic over
the backend in the future.