Remove `QueryBuilder` related methods from `JoinTo`
This is part of an ongoing refactoring to consolidate all `to_sql` type
methods on `QueryFragment`

This code gets a little bit messy in places, particularly in keeping
`join_through!` around. This will eventually go away, but there is some
interesting bits in the `CombineJoin` trait which makes me think we
might be able to derive `join_through` in interesting ways.

The code will likly get a little bit cleaner once I do this same
refactoring to `QuerySource`.

I beleive that with this change we can actually remove `InnerJoinSource`
and `LeftOuterJoinSource` but I haven't explored it just yet.