Remove the lifetime from the `PgQueryBuilder`
As we work to make `QueryFragment` generic over the backend, I would
prefer to not have a lifetime on the `Pg` backend. As such, that means
that `PgQueryBuilder` needs to have no associated lifetime. I don't want
to require that you put the actual connection in an `Rc`, but I do need
to pass something that isn't a reference to the query builder.

As such, I've isolated all methods that work with PQ directly into
another structure, which we keep inside of the `Connection`. This does
introduce a potential subtlety, which is that we are `Send` even though
we have a non-atomic reference counter. The reason this is safe is
because we *create* the `Rc` internally, and the only place we hand it
to anybody else is for a `query_builder`, which is only created in a
place that does not outlive a reference to the `Connection`. That means
that in order to send the `Connection`, you'd have to be in a context
where there are no other references to the `RawConnection`. However,
this is no longer enforced by the type system.