Turn `Connection` into a trait
This was pretty straightforward. Almost everything that was previously
coupled to PG has been changed to be generic over the connection (mainly
`LoadDsl` and `ExecuteDsl`). Our tests are coupled to PG. We'll want to
change that as we further break up PG into it's own crate. Things that
apply to Sqlite will need to be tested there as well.

`SimpleConnection` exists because I didn't want to make migrations
generic over the connection, but `Connection` is not object safe. We
only ever use `batch_execute` in that module, so I'm just taking that as
a trait object.

CLI and Codegen are currently coupled to PG. I will need to write
additional code for both of them in order to support SQLite in the
future. It occurs to me that I don't have a great way to make these
support third party adapters, and I'll need to rethink how the
connection info is passed in.