Kill the hellish demon that is `Connection#find`
Let's be honest here. Nobody has seen a where clause in Rust than this.
It has become ridiculous, and in the past week, just about every
substantial commit I've made has caused it to become more ridiculous.

This is the *only* method on `Connection` not related to transactions
(which -- correct me if I'm wrong -- will always need to be methods on
`Connection`). This came out of a time where `query.load(&connection)`
was `connection.query_all(&query)`. This method is inconsistent, and in
our documentation is just ridiculous. Please, let's kill it.

The replacement for this method does not implicitly add `limit(1)`. If
you call `.get_result`, instead of `.first`, there's no actual guarantee
that the result set it's operating on contains only a single row.
However, given that this method filters on equality for a primary key, I
think this is low risk.

I honestly couldn't remember why I added this method at all. For those
curious, it was added in 77f2e39, and the reasoning was "It was easier
to implement than `filter` at the time". This was commit number 42,
which clearly means it is in fact the meaning of Rust, Diesel, and
everything, but it's time for it to go.

Also the original implementation included `let sql = sql + &format!("
WHERE {} = $1 LIMIT 1", source.primary_key().name());` which I find
completely hilarious. Ah ancient Diesel, your code was once so bad...