Make sure we only use `RETURNING` on backends that support it
SQLite does not support RETURNING clauses. To my knowledge it is the
only backend that does not support it (though the syntax is different on
SQL Server).

Of course at the moment the only backend supported on master is PG,
which *does* support it. However, this was an extraction from the SQLite
branch that is able to live on its own. I've removed most uses of
`get_result` and `get_results` in tests, except for one explicit test
that returning clauses work (which is currently in a cfg attr which is
always true, but won't be once SQLite is merged).

There are still some places that it's being used which I need to track
down, but right now I've got those modules set to only run on PG anyway,
as they have other issues that require detangling from PG.