Setup travis in a way that doesn't use ancient SQLite versions
The version that is in the container infrastructure is from 2011. We use
`sqlite3_errstr` which was added in 2012. The switch to the "trusty"
infratstructure bumps it to 3.8, which is relatively recent.

This exposes a legitimate bug related to nul bytes in string fields that
for some reason not replicating for me locally.