Clean up the migration signatures to something reasonable
The where clauses that were needed when we moved to a generic backend
structure are leaking into the docs, making these functions seem way
more scary than they actually are (this could just as easily be
`&Connection` if connection were object safe).

I've moved everything onto another trait so that the signatures of these
functions look a bit more reasonable.